backend/requirements.txt
graphrag==3.0.8
PyMuPDF>=1.27
beautifulsoup4>=4.12
requests>=2.31
playwright>=1.55
flask>=3.0
flask-cors>=5.0
networkx>=3.2
numpy>=1.26
python-dotenv>=1.0
